Move data mapper classes to BaseStructures.py
