+++ Assets/Python/data/maps/SettlerMapData.py
@@ -1,4 +1,4 @@
-from CoreStructures import CivDataMapper
+from BaseStructures import CivDataMapper
 from CoreTypes import Civ
 
 SETTLERS_MAP = CivDataMapper({
